combo <none> kanban <other> column refreshing issues navigator summary

+++ src/list.board.move.svelte
@@ -5,8 +5,8 @@
     export let taskStates;
     export let item;
     export let afterActionOperation = undefined;
-    export let onMoveUp;
-    export let onMoveDown;
+    export let onMoveUp = undefined;
+    export let onMoveDown = undefined;
     export let onReplace;
 
     function replaceMenu()
@@ -20,26 +20,28 @@
 <!-- svelte-ignore a11y-click-events-have-key-events -->
 
 <menu class="hidden sm:flex flex-row gap-3">
-    <section class="flex flex-col gap-1">
-        <p class="text-xs h-4 w-[100px] truncate"></p>
-        <button class=" h-8 w-full
-                        text-xs font-medium text-stone-900 dark:text-stone-400 dark:hover:text-white
-                        bg-stone-100 hover:bg-stone-200 dark:bg-stone-800 dark:hover:bg-stone-700 active:bg-stone-300 dark:active:bg-stone-600
-                        border rounded border-stone-200 dark:border-stone-600 focus:outline-none
-                        inline-flex items-center justify-center"
-                        on:click={(e) => { onMoveUp(item); replaceMenu()}}>
-            <div class="w-3 h-3"><FaAngleUp/></div>
-        </button>
+    {#if onMoveUp && onMoveDown}
+        <section class="flex flex-col gap-1">
+            <p class="text-xs h-4 w-[100px] truncate"></p>
+            <button class=" h-8 w-full
+                            text-xs font-medium text-stone-900 dark:text-stone-400 dark:hover:text-white
+                            bg-stone-100 hover:bg-stone-200 dark:bg-stone-800 dark:hover:bg-stone-700 active:bg-stone-300 dark:active:bg-stone-600
+                            border rounded border-stone-200 dark:border-stone-600 focus:outline-none
+                            inline-flex items-center justify-center"
+                            on:click={(e) => { onMoveUp(item); replaceMenu()}}>
+                <div class="w-3 h-3"><FaAngleUp/></div>
+            </button>
 
-        <button class=" h-8 w-full
-                        text-xs font-medium text-stone-900 dark:text-stone-400 dark:hover:text-white
-                        bg-stone-100 hover:bg-stone-200 dark:bg-stone-800 dark:hover:bg-stone-700 active:bg-stone-300 dark:active:bg-stone-600
-                        border rounded border-stone-200 dark:border-stone-600 focus:outline-none
-                        inline-flex items-center justify-center"
-                        on:click={(e) => { onMoveDown(item); replaceMenu()}}>
-            <div class="w-3 h-3"><FaAngleDown/></div>
-        </button>
-    </section>
+            <button class=" h-8 w-full
+                            text-xs font-medium text-stone-900 dark:text-stone-400 dark:hover:text-white
+                            bg-stone-100 hover:bg-stone-200 dark:bg-stone-800 dark:hover:bg-stone-700 active:bg-stone-300 dark:active:bg-stone-600
+                            border rounded border-stone-200 dark:border-stone-600 focus:outline-none
+                            inline-flex items-center justify-center"
+                            on:click={(e) => { onMoveDown(item); replaceMenu()}}>
+                <div class="w-3 h-3"><FaAngleDown/></div>
+            </button>
+        </section>
+    {/if}
 
     {#if taskStates && taskStates.length > 0}
     {#each taskStates as taskState, idx}
@@ -68,44 +70,46 @@
 </menu>
 
 <menu class="flex sm:hidden flex-col gap-3">
-    <section class="flex flex-row">
-        <p class="text-base truncate "></p>
-        <div class="flex flex-row ml-auto gap-3">
-            <button class=" bg-transparent
-                            w-[55px] h-[55px] 
-                            m-0
-                            flex items-center justify-center
-                            disable-dbl-tap-zoom"
-                            on:click={(e) => { onMoveUp(item); replaceMenu()}}>
-                <div class="    w-10 h-10
-                                text-white bg-zinc-500 hover:bg-zinc-500 
-                                font-medium rounded-full text-sm text-center shadow-md
-                                dark:bg-zinc-500 dark:hover:bg-zinc-500
-                                flex items-center justify-center">
-                    <div class="w-5 h-5">
-                        <FaAngleUp/>
+    {#if onMoveUp && onMoveDown}
+        <section class="flex flex-row">
+            <p class="text-base truncate "></p>
+            <div class="flex flex-row ml-auto gap-3">
+                <button class=" bg-transparent
+                                w-[55px] h-[55px] 
+                                m-0
+                                flex items-center justify-center
+                                disable-dbl-tap-zoom"
+                                on:click={(e) => { onMoveUp(item); replaceMenu()}}>
+                    <div class="    w-10 h-10
+                                    text-white bg-zinc-500 hover:bg-zinc-500 
+                                    font-medium rounded-full text-sm text-center shadow-md
+                                    dark:bg-zinc-500 dark:hover:bg-zinc-500
+                                    flex items-center justify-center">
+                        <div class="w-5 h-5">
+                            <FaAngleUp/>
+                        </div>
                     </div>
-                </div>
-            </button>
+                </button>
 
-            <button class=" bg-transparent
-                            w-[55px] h-[55px] 
-                            m-0
-                            flex items-center justify-center
-                            disable-dbl-tap-zoom"
-                            on:click={(e) => { onMoveDown(item); replaceMenu()}}>
-                <div class="    w-10 h-10
-                                text-white bg-zinc-500 hover:bg-zinc-500 
-                                font-medium rounded-full text-sm text-center shadow-md
-                                dark:bg-zinc-500 dark:hover:bg-zinc-500
-                                flex items-center justify-center">
-                    <div class="w-5 h-5">
-                        <FaAngleDown/>
+                <button class=" bg-transparent
+                                w-[55px] h-[55px] 
+                                m-0
+                                flex items-center justify-center
+                                disable-dbl-tap-zoom"
+                                on:click={(e) => { onMoveDown(item); replaceMenu()}}>
+                    <div class="    w-10 h-10
+                                    text-white bg-zinc-500 hover:bg-zinc-500 
+                                    font-medium rounded-full text-sm text-center shadow-md
+                                    dark:bg-zinc-500 dark:hover:bg-zinc-500
+                                    flex items-center justify-center">
+                        <div class="w-5 h-5">
+                            <FaAngleDown/>
+                        </div>
                     </div>
-                </div>
-            </button>
-        </div>
-    </section>
+                </button>
+            </div>
+        </section>
+    {/if}
 
     {#if taskStates && taskStates.length > 0}
     {#each taskStates as taskState, idx}
